+++ src/gufo/thor/docker.py
@@ -144,6 +144,14 @@ class Docker(object):
         logger.warning("Stopping containers")
         return self._commpose_command("stop", _exec=True)
 
+    def logs(self: "Docker", *args: str, _follow: bool = False) -> bool:
+        """Show logs."""
+        cmd = ["logs"]
+        if _follow:
+            cmd.append("-f")
+        cmd.extend(args)
+        return self._commpose_command(*cmd, _exec=True)
+
     def restart(self: "Docker", *args: str) -> bool:
         """
         Perform services restart.